backend/requirements.txt
fastapi==0.119.1
pydantic==2.12.3
starlette==0.48.0
uvicorn==0.38.0
python-multipart==0.0.20
openai-whisper==20250625
ffmpeg-python==0.2.0
celery==5.5.3
redis==6.4.0
python-dotenv==1.1.1
pyannote-audio==4.0.1
torchcodec==0.8.0
reportlab==4.4.4
python-docx==1.2.0
yt-dlp==2025.10.22